+++ RecipeManager-base/src/main/java/haveric/recipeManager/flag/flags/result/FlagFood.java
@@ -1,10 +1,18 @@
 package haveric.recipeManager.flag.flags.result;
 
+import com.google.common.collect.ObjectArrays;
 import haveric.recipeManager.ErrorReporter;
 import haveric.recipeManager.Files;
 import haveric.recipeManager.flag.Flag;
 import haveric.recipeManager.flag.FlagType;
+import haveric.recipeManager.flag.args.ArgBuilder;
 import haveric.recipeManager.flag.args.Args;
+import haveric.recipeManager.recipes.ItemResult;
+import haveric.recipeManager.recipes.item.ItemRecipe;
+import haveric.recipeManager.tools.Tools;
+import haveric.recipeManager.tools.Version;
+import org.bukkit.Material;
+import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.meta.ItemMeta;
 import org.bukkit.inventory.meta.components.FoodComponent;
 
@@ -23,20 +31,34 @@ public class FlagFood extends Flag {
 
     @Override
     protected String[] getDescription() {
-        return new String[] {
+        String[] description = new String[]{
             "Adds a food component to an item, making it consumable",
             "Using this flag more than once will overwrite the previous flag.",
             "",
             "  The <nutrition> argument sets food restored by this item when eaten. Value must be a non-negative integer. Defaults to 0.",
             "",
             "Optionally you can specify some arguments separated by | character:",
-            "  saturation <float>  = (default 0.0) saturation restored by this item when eaten. Value must be float value.",
-            "  alwayseat [false]   = (default false) if set, this item can be eaten even when not hungry.",
-            "  seconds <float>     = (default 0.0) time in seconds it will take for this item to be eaten.",
+            "  saturation <float> = (default 0.0) saturation restored by this item when eaten. Value must be float value.",
+            "  alwayseat [false] = (default false) if set, this item can be eaten even when not hungry.",
+            "  seconds <float> = (default 0.0) time in seconds it will take for this item to be eaten.",
             "    NOTE: The default value of 0.0 will not consume the item!",
+        };
+
+        if (Version.has1_21Support()) {
+            description = ObjectArrays.concat(description, new String[]{
+                "  convertsto <item>[:data][:amount] = sets the item this food will convert to once eaten.",
+                "  convertsto item:<name> = sets the item this food will convert to once eaten.",
+                "    <name> = The name of an item recipe defined before this flag."
+            }, String.class);
+        }
+
+        description = ObjectArrays.concat(description, new String[]{
             "You can specify these arguments in any order and they're completely optional.",
             "",
-            "NOTE: You can also add Potion effects with " + FlagType.FOOD_POTION_EFFECT, };
+            "NOTE: You can also add Potion effects with " + FlagType.FOOD_POTION_EFFECT,
+        }, String.class);
+
+        return description;
     }
 
     @Override
@@ -50,6 +72,7 @@ public class FlagFood extends Flag {
     private float saturation = 0.0f;
     private boolean alwaysEat = false;
     private float seconds = 0.0f;
+    private ItemRecipe convertsToItemRecipe = null;
 
     public FlagFood() {
 
@@ -61,11 +84,17 @@ public class FlagFood extends Flag {
         saturation = flag.saturation;
         alwaysEat = flag.alwaysEat;
         seconds = flag.seconds;
+        convertsToItemRecipe = flag.convertsToItemRecipe;
     }
 
     @Override
     public FlagFood clone() {
         return new FlagFood((FlagFood) super.clone());
+    }
+
+    @Override
+    public boolean requiresRecipeManagerModification() {
+        return convertsToItemRecipe == null;
     }
 
     @Override
@@ -109,6 +138,28 @@ public class FlagFood extends Flag {
                     } catch (IllegalArgumentException e) {
                         return ErrorReporter.getInstance().error("Flag " + getFlagType() + " has invalid seconds value: " + value);
                     }
+                } else if (Version.has1_21Support() && value.startsWith("convertsto")) {
+                    value = value.substring("convertsto".length()).trim();
+
+                    if (value.startsWith("item:")) {
+                        value = value.substring("item:".length());
+
+                        ItemRecipe recipe = ItemRecipe.getRecipe(value);
+                        if (recipe == null) {
+                            return ErrorReporter.getInstance().error("Flag " + getFlagType() + " has invalid item reference: " + value + "!");
+                        } else {
+                            convertsToItemRecipe = recipe;
+                        }
+                    } else {
+                        ItemStack item = Tools.parseItem(value, 0);
+                        if (item == null || item.getType() == Material.AIR) {
+                            return ErrorReporter.getInstance().error("Flag " + getFlagType() + " has invalid item defined!");
+                        } else {
+                            ItemRecipe recipe = new ItemRecipe();
+                            recipe.setResult(item);
+                            convertsToItemRecipe = recipe;
+                        }
+                    }
                 } else {
                     ErrorReporter.getInstance().warning("Flag " + getFlagType() + " has unknown argument: " + value, "Maybe it's spelled wrong, check it in " + Files.FILE_INFO_FLAGS + " file.");
                 }
@@ -135,6 +186,16 @@ public class FlagFood extends Flag {
                 food.setSaturation(saturation);
                 food.setCanAlwaysEat(alwaysEat);
                 food.setEatSeconds(seconds);
+                if (Version.has1_21Support()) {
+                    if (convertsToItemRecipe != null) {
+                        ItemResult result = convertsToItemRecipe.getResult();
+                        Args itemArgs = ArgBuilder.create(a).recipe(convertsToItemRecipe).result(result).build();
+                        itemArgs.setFirstRun(true);
+                        if (result.getFlags().sendCrafted(itemArgs)) {
+                            food.setUsingConvertsTo(itemArgs.result());
+                        }
+                    }
+                }
 
                 meta.setFood(food);
             }