+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Change log
 
 ### v2.31.1
+* NEW: `@Food` convertsto argument to change food once eaten
 * FIXED: 1.20.4-: `@IngredientCondition`, `@HoldItem` NoClassDefFoundError for ItemRarity
 * FIXED: 1.20.1-: `@PotionItem` NoClassDefFoundError for getBasePotionType
 * RE-ADD: 1.20.1-: `@PotionItem` arguments: `level`, `extended`